running add to favourites feature
