+++ .idea/misc.xml
@@ -3,6 +3,7 @@
   <component name="DesignSurface">
     <option name="filePathToZoomLevelMap">
       <map>
+        <entry key="../../../../layout/custom_preview.xml" value="0.1" />
         <entry key="app/src/main/res/drawable-v24/ic_launcher_foreground.xml" value="0.321025641025641" />
         <entry key="app/src/main/res/drawable/add_icon.xml" value="0.18888888888888888" />
         <entry key="app/src/main/res/drawable/ic_broken_image.xml" value="0.1" />
@@ -10,15 +11,17 @@
         <entry key="app/src/main/res/drawable/ic_launcher_background.xml" value="0.321025641025641" />
         <entry key="app/src/main/res/drawable/loading_animation.xml" value="0.1" />
         <entry key="app/src/main/res/drawable/loading_img.xml" value="0.1" />
+        <entry key="app/src/main/res/layout/activity_login.xml" value="0.1" />
         <entry key="app/src/main/res/layout/activity_main.xml" value="0.1" />
+        <entry key="app/src/main/res/layout/activity_splash_screen.xml" value="0.1023772652458393" />
         <entry key="app/src/main/res/layout/fragment_favorites.xml" value="0.1" />
-        <entry key="app/src/main/res/layout/fragment_home.xml" value="0.14814814814814814" />
+        <entry key="app/src/main/res/layout/fragment_home.xml" value="0.1" />
         <entry key="app/src/main/res/layout/fragment_home_page.xml" value="0.1" />
         <entry key="app/src/main/res/layout/fragment_login.xml" value="0.1" />
         <entry key="app/src/main/res/layout/fragment_over_view.xml" value="0.1" />
-        <entry key="app/src/main/res/layout/fragment_travel_plans.xml" value="0.30512820512820515" />
+        <entry key="app/src/main/res/layout/fragment_travel_plans.xml" value="0.1" />
         <entry key="app/src/main/res/layout/grid_view_item.xml" value="0.1" />
-        <entry key="app/src/main/res/layout/list_item.xml" value="0.2740740740740741" />
+        <entry key="app/src/main/res/layout/list_item.xml" value="0.375" />
         <entry key="app/src/main/res/menu/bottom_app_bar.xml" value="0.1" />
         <entry key="app/src/main/res/menu/bottom_navigation_menu.xml" value="0.30512820512820515" />
         <entry key="app/src/main/res/menu/main_menu.xml" value="0.3098958333333333" />
@@ -30,5 +33,12 @@
   </component>
   <component name="ProjectType">
     <option name="id" value="Android" />
+  </component>
+  <component name="VisualizationToolProject">
+    <option name="state">
+      <ProjectState>
+        <option name="scale" value="0.1" />
+      </ProjectState>
+    </option>
   </component>
 </project>
