+++ app/src/main/java/com/raghad/goexplore/MainActivity.kt
@@ -1,5 +1,6 @@
 package com.raghad.goexplore
 
+import android.content.Intent
 import androidx.appcompat.app.AppCompatActivity
 import android.os.Bundle
 import android.view.Menu
@@ -7,25 +8,30 @@ import android.view.MenuItem
 import androidx.navigation.NavController
 import androidx.navigation.findNavController
 import androidx.navigation.fragment.NavHostFragment
-import androidx.navigation.fragment.findNavController
 import androidx.navigation.ui.setupActionBarWithNavController
+import androidx.navigation.ui.setupWithNavController
+import com.google.android.material.bottomnavigation.BottomNavigationView
 import com.google.firebase.auth.FirebaseAuth
-import com.raghad.goexplore.pages.HomeFragmentDirections
 
 class MainActivity : AppCompatActivity() {
 
     private lateinit var navController: NavController
-
 
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
         setContentView(R.layout.activity_main)
 
         val navHostFragment = supportFragmentManager
-            .findFragmentById(R.id.overviewFragment) as NavHostFragment
+            .findFragmentById(R.id.nav_host_fragment) as NavHostFragment
         navController = navHostFragment.navController
 
         setupActionBarWithNavController(navController)
+
+        val bottomNavigationView = findViewById<BottomNavigationView>(R.id.bottom_navigation)
+        val navController = findNavController(R.id.nav_host_fragment)
+
+        bottomNavigationView.setupWithNavController(navController)
+
     }
 
     override fun onCreateOptionsMenu(menu: Menu?): Boolean {
@@ -40,14 +46,10 @@ class MainActivity : AppCompatActivity() {
         if(id == R.id.logout){
 
             FirebaseAuth.getInstance().signOut()
+
+            val intent = Intent(this, LoginActivity::class.java)
             startActivity(intent)
-
-            var action = HomeFragmentDirections.actionHomeFragmentToLoginFragment()
-            findNavController(R.id.overviewFragment).navigate(action)
-
         }
         return super.onOptionsItemSelected(item)
-
     }
-
 }
