+++ app/src/main/java/com/raghad/goexplore/model/Favourites.kt
@@ -1,0 +1,5 @@
+package com.raghad.goexplore.model
+
+data class Favourites(val favouriteItem : String) {
+
+}
