+++ app/src/main/java/com/raghad/goexplore/ui/FavoritesFragment.kt
@@ -1,0 +1,32 @@
+package com.raghad.goexplore.ui
+
+import android.os.Bundle
+import androidx.fragment.app.Fragment
+import android.view.LayoutInflater
+import android.view.View
+import android.view.ViewGroup
+import com.raghad.goexplore.R
+
+
+class FavoritesFragment : Fragment() {
+
+    override fun onCreate(savedInstanceState: Bundle?) {
+        super.onCreate(savedInstanceState)
+        arguments?.let {
+
+        }
+    }
+
+    override fun onCreateView(
+        inflater: LayoutInflater, container: ViewGroup?,
+        savedInstanceState: Bundle?
+    ): View? {
+        // Inflate the layout for this fragment
+        return inflater.inflate(R.layout.fragment_favorites, container, false)
+    }
+
+
+
+    companion object {
+    }
+}
