+++ app/src/main/java/com/raghad/goexplore/ui/HomeFragment.kt
@@ -1,17 +1,11 @@
-package com.raghad.goexplore.pages
+package com.raghad.goexplore.ui
 
-import android.content.Intent
 import android.os.Bundle
 import android.view.*
 import androidx.fragment.app.Fragment
 import androidx.fragment.app.viewModels
-import androidx.navigation.fragment.findNavController
-import com.google.android.material.bottomnavigation.BottomNavigationView
-import com.google.firebase.auth.FirebaseAuth
-import com.raghad.goexplore.R
 import com.raghad.goexplore.adapter.PhotoGridAdapter
 import com.raghad.goexplore.databinding.FragmentHomeBinding
-import com.raghad.goexplore.databinding.FragmentLoginBinding
 import com.raghad.goexplore.overview.OverViewViewModel
 
 /*
