+++ app/src/main/res/layout/activity_login.xml
@@ -1,12 +1,11 @@
 <?xml version="1.0" encoding="utf-8"?>
-
 <androidx.constraintlayout.widget.ConstraintLayout xmlns:android="http://schemas.android.com/apk/res/android"
     xmlns:app="http://schemas.android.com/apk/res-auto"
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".MainActivity"
-    android:background="@color/white">
+    tools:context=".LoginActivity">
+
 
     <Button
         android:id="@+id/Button"
@@ -23,24 +22,6 @@
         app:layout_constraintStart_toStartOf="parent"
         app:layout_constraintTop_toTopOf="parent" />
 
-
-<!--    <Button-->
-<!--        android:id="@+id/home"-->
-<!--        android:layout_width="160dp"-->
-<!--        android:layout_height="60dp"-->
-<!--        android:layout_marginStart="125dp"-->
-<!--        android:layout_marginTop="500dp"-->
-<!--        android:layout_marginEnd="125dp"-->
-<!--        android:layout_marginBottom="80dp"-->
-<!--        android:gravity="center"-->
-<!--        android:text="home"-->
-<!--        app:layout_constraintBottom_toBottomOf="parent"-->
-<!--        app:layout_constraintEnd_toEndOf="parent"-->
-<!--        app:layout_constraintHorizontal_bias="1.0"-->
-<!--        app:layout_constraintStart_toStartOf="parent"-->
-<!--        app:layout_constraintTop_toTopOf="parent"-->
-<!--        app:layout_constraintVertical_bias="0.0" />-->
-
     <com.airbnb.lottie.LottieAnimationView
         android:id="@+id/animationView"
         android:layout_width="331dp"
@@ -55,5 +36,8 @@
         app:lottie_loop="true"
         app:lottie_rawRes="@raw/animation" />
 
-
 </androidx.constraintlayout.widget.ConstraintLayout>
+
+
+
+
