+++ app/src/main/res/layout/grid_view_item.xml
@@ -16,9 +16,9 @@
     <ImageView
         android:id="@+id/place_image"
         android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
+        android:layout_height="180dp"
         android:adjustViewBounds="true"
-        android:padding="2dp"
+        android:padding="4dp"
         android:scaleType="fitXY"
         app:imageUrl="@{photo.imageUrl}"
         tools:src="@tools:sample/backgrounds/scenic"
