+++ app/src/main/res/navigation/navigation_graph.xml
@@ -3,19 +3,11 @@
     xmlns:app="http://schemas.android.com/apk/res-auto"
     xmlns:tools="http://schemas.android.com/tools"
     android:id="@+id/navigation_graph"
-    app:startDestination="@id/loginFragment">
+    app:startDestination="@id/homeFragment">
 
     <fragment
-        android:id="@+id/loginFragment"
-        android:name="com.raghad.goexplore.pages.LoginFragment"
-        android:label="Go Explore">
-        <action
-            android:id="@+id/action_loginFragment_to_homeFragment"
-            app:destination="@id/homeFragment" />
-    </fragment>
-    <fragment
         android:id="@+id/homeFragment"
-        android:name="com.raghad.goexplore.pages.HomeFragment"
+        android:name="com.raghad.goexplore.ui.HomeFragment"
         android:label="Start Explore">
         <action
             android:id="@+id/action_homeFragment_to_travelPlansFragment"
@@ -23,14 +15,12 @@
         <action
             android:id="@+id/action_homeFragment_to_favoritesFragment"
             app:destination="@id/favoritesFragment" />
-        <action
-            android:id="@+id/action_homeFragment_to_loginFragment"
-            app:destination="@id/loginFragment" />
     </fragment>
+
     <fragment
         android:id="@+id/favoritesFragment"
-        android:name="com.raghad.goexplore.pages.FavoritesFragment"
-        android:label="FFF"
+        android:name="com.raghad.goexplore.ui.FavoritesFragment"
+        android:label="Favourites"
         tools:layout="@layout/fragment_favorites">
         <action
             android:id="@+id/action_favoritesFragment_to_travelPlansFragment"
@@ -39,11 +29,12 @@
             android:id="@+id/action_favoritesFragment_to_homeFragment"
             app:destination="@id/homeFragment" />
     </fragment>
+
     <fragment
         android:id="@+id/travelPlansFragment"
-        android:name="com.raghad.goexplore.pages.TravelPlansFragment"
-        android:label="fragment_travel_plans"
-        tools:layout="@layout/fragment_travel_plans" >
+        android:name="com.raghad.goexplore.ui.TravelPlansFragment"
+        android:label="Plans"
+        tools:layout="@layout/fragment_travel_plans">
         <action
             android:id="@+id/action_travelPlansFragment_to_homeFragment"
             app:destination="@id/homeFragment" />
@@ -51,5 +42,14 @@
             android:id="@+id/action_travelPlansFragment_to_favoritesFragment"
             app:destination="@id/favoritesFragment" />
     </fragment>
+
+    <activity
+        android:id="@+id/mainActivity"
+        android:name="com.raghad.goexplore.MainActivity" />
+    <activity
+        android:id="@+id/loginActivity"
+        android:name="com.raghad.goexplore.LoginActivity"
+        android:label="LoginActivity"
+        tools:layout="@layout/fragment_login"/>
 
 </navigation>
