+++ app/src/main/res/values/strings.xml
@@ -7,4 +7,5 @@
     <string name="plan">Plan</string>
     <string name="explore">Explore</string>
     <string name="favorites">Favorites</string>
+    <string name="favourites">Favourites</string>
 </resources>
